time-tool: added flag "show records" for sync command

+++ apps/time-tool/cmd/sync.go
@@ -7,6 +7,8 @@ import (
 	"github.com/spf13/cobra"
 )
 
+var FlagSyncShow = false
+
 // Synchronizing local storage with the server
 var SyncCmd = &cobra.Command{
 	Use:   "sync",
@@ -14,14 +16,18 @@ var SyncCmd = &cobra.Command{
 	Run: func(cmd *cobra.Command, args []string) {
 		timestamps, _ := api.GetTimestamps()
 		for _, timestamp := range timestamps {
-			timestampRes, err := api.PushTimestamp(timestamp)
-			if err != nil {
-				core.Danger("Error: %v\n", err.Error())
-				return
+			if !FlagSyncShow {
+				_, err := api.PushTimestamp(timestamp)
+				if err != nil {
+					core.Danger("Error: %v\n", err.Error())
+					return
+				}
 			}
-			core.Info("Timestamp (%s): %s\n", timestampRes.Data.Type, timestampRes.Data.Timestamp)
+			core.Info("Timestamp (%s): %s\n", timestamp.Type, timestamp.Timestamp)
 		}
-		api.DeleteTimestamps()
-		core.Success("Your data is synced!\n")
+		if !FlagSyncShow {
+			api.DeleteTimestamps()
+			core.Success("Your data is synced!\n")
+		}
 	},
 }
